+++ backend/windmill-worker/src/worker_flow.rs
@@ -425,29 +425,18 @@ pub async fn update_flow_status_after_job_completion_internal<
                     tx.commit().await?;
 
                     if parallelism.is_some() {
-                        // this ensure that the lock is taken in the same order and thus avoid deadlocks
-                        let ids = sqlx::query_scalar!(
-                            "SELECT id FROM queue WHERE parent_job = $1 AND suspend > 0 ORDER by suspend",
-                            flow
+                        sqlx::query!(
+                            "UPDATE queue SET suspend = 0 WHERE parent_job = $1 AND suspend = $2",
+                            flow,
+                            nindex
                         )
-                        .fetch_all(db)
+                        .execute(db)
                         .await
                         .map_err(|e| {
-                            Error::InternalErr(format!("error while locking jobs to decrease parallelism of: {e:#}"))
+                            Error::InternalErr(format!(
+                                "error resuming job at suspend {nindex} and parent {flow}: {e:#}"
+                            ))
                         })?;
-                        for id in ids {
-                            sqlx::query!(
-                                "UPDATE queue SET suspend = suspend - 1 WHERE id = $1 AND suspend > 0",
-                                id
-                            )
-                            .execute(db)
-                            .await
-                            .map_err(|e| {
-                                Error::InternalErr(format!(
-                                    "error decreasing suspend for {id}: {e:#}"
-                                ))
-                            })?;
-                        }
                     }
 
                     sqlx::query!(
@@ -1986,6 +1975,18 @@ async fn push_next_flow_job<R: rsmq_async::RsmqConnection + Send + Sync + Clone>
     let mut tx: QueueTransaction<'_, R> = (rsmq.clone(), db.begin().await?).into();
 
     for i in (0..len).into_iter() {
+        if i % 100 == 0 && i != 0 {
+            tracing::info!(id = %flow_job.id, root_id = %job_root, "pushed (non-commited yet) first {i} subflows of {len}");
+            sqlx::query!(
+                "UPDATE queue
+                SET last_ping = now()
+                WHERE id = $1 AND last_ping < now()",
+                flow_job.id,
+            )
+            .execute(db)
+            .await?;
+        }
+        tracing::debug!(id = %flow_job.id, root_id = %job_root, "pushing job {i} of {len}");
         let payload_tag = match &job_payloads {
             ContinuePayload::SingleJob(payload) => payload.clone(),
             ContinuePayload::BranchAllJobs(payloads) => payloads[i].clone(),
@@ -2085,6 +2086,8 @@ async fn push_next_flow_job<R: rsmq_async::RsmqConnection + Send + Sync + Clone>
             Err(e) => (None, Some(e)),
         };
 
+        tracing::debug!(id = %flow_job.id, root_id = %job_root, "computed args for job {i} of {len}");
+
         let value_with_parallel = module.get_value_with_parallel()?;
 
         let root_job = if {
@@ -2115,6 +2118,8 @@ async fn push_next_flow_job<R: rsmq_async::RsmqConnection + Send + Sync + Clone>
         } else {
             None
         };
+
+        tracing::debug!(id = %flow_job.id, root_id = %job_root, "computed perms for job {i} of {len}");
 
         let tx2 = PushIsolationLevel::Transaction(tx);
         let (uuid, mut inner_tx) = push(
@@ -2154,6 +2159,8 @@ async fn push_next_flow_job<R: rsmq_async::RsmqConnection + Send + Sync + Clone>
 
         if value_with_parallel.type_ == "forloopflow" {
             if let Some(p) = value_with_parallel.parallelism {
+                tracing::debug!(id = %flow_job.id, root_id = %job_root, "updating suspend for forloopflow job {uuid}");
+
                 if i as u16 >= p {
                     sqlx::query!(
                         "UPDATE queue
@@ -2165,6 +2172,7 @@ async fn push_next_flow_job<R: rsmq_async::RsmqConnection + Send + Sync + Clone>
                     .execute(&mut inner_tx)
                     .await?;
                 }
+                tracing::debug!(id = %flow_job.id, root_id = %job_root, "updated suspend for {uuid}");
             }
         }
 
@@ -2206,6 +2214,7 @@ async fn push_next_flow_job<R: rsmq_async::RsmqConnection + Send + Sync + Clone>
             )
             .execute(&mut tx)
             .await?;
+            tracing::debug!(id = %flow_job.id, root_id = %job_root, "updated parallel monitor lock for {uuid}");
         }
     }
     let first_uuid = uuids[0];